Remove Redundant Casts (IntelliJ quick-fix applied to full project)

+++ aspose/src/main/java/org/frankframework/extensions/aspose/services/CisConversionResult.java
@@ -87,7 +87,7 @@ public class CisConversionResult {
 				}
 			}
 		} else {
-			numberOfPages = (int) message.getContext().get("Pdf.Pages");
+			numberOfPages = message.getContext().get("Pdf.Pages");
 		}
 	}
